fix(validation): close Copilot round-2 findings in the citation gate
Four behavior fixes, each pinned by a regression test committed
alongside:

- URLs are masked before the citation scan, so an external link shaped
  like host/org/file.py:N never reads as a repository citation.
- A slashless bare filename is masked before identifier harvest, so its
  stem is no longer a required anchor for an unrelated citation.
- Anchors on a multi-citation line bind to their nearer citation via a
  midpoint split snapped so it never bisects a span, quote, or
  identifier; neighbor lines stay shared because a wrapped sentence's
  anchor legitimately serves every citation it describes. Replaying PR
  #5327's merged diff still re-finds exactly the three stale citations
  open PR #5336 corrects.
- The diff parse pins git's deterministic textual flags, matching
  scripts/validation/git_hook_policy.py:274 verbatim
  (TEXTUAL_DIFF_FLAGS = ("--no-ext-diff", "--no-textconv", "--text")),
  so a contributor's markdown textconv driver cannot rewrite the patch
  the parser reads.

Anchor semantics moved to citation_anchors.py at the file-size ceiling,
completing the seam citation_head_state.py started: git reads, anchor
semantics, and gate policy now live in one module each.

Refs #5337
